[FEAT] #221 탭 패딩값 조절 및 태그 UI 구현
