+++ Whistle/UI/Components/Content/BookmarkedContentLayer.swift
@@ -141,7 +141,7 @@ struct BookmarkedContentLayer: View {
               }
             } label: {
               ContentLayerButton(
-                type: .whistle(currentVideoInfo.whistleCount),
+                type: .whistle(currentVideoInfo.whistleCount.roundedWithAbbreviations),
                 isFilled: $currentVideoInfo.isWhistled)
 //              ContentLayerButton(
 //                isFilled: $currentVideoInfo.isWhistled,
