+++ Whistle/UI/Components/Content/MainContentLayer.swift
@@ -139,7 +139,7 @@ struct MainContentLayer: View {
               }
             } label: {
               ContentLayerButton(
-                type: .whistle(currentVideoInfo.whistleCount),
+                type: .whistle(currentVideoInfo.whistleCount.roundedWithAbbreviations),
                 isFilled: $currentVideoInfo.isWhistled)
             }
             .buttonStyle(PressEffectButtonStyle())
