+++ Whistle/UI/Components/Content/MemberContentLayer.swift
@@ -135,7 +135,7 @@ struct MemberContentLayer: View {
               }
             } label: {
               ContentLayerButton(
-                type: .whistle(currentVideoInfo.whistleCount),
+                type: .whistle(currentVideoInfo.whistleCount.roundedWithAbbreviations),
                 isFilled: $currentVideoInfo.isWhistled)
             }
             Button {
