+++ Whistle/UI/Screens/Search/MainSearchView.swift
@@ -81,94 +81,3 @@ struct MainSearchView: View {
     }
   }
 }
-
-#Preview {
-  MainSearchView()
-}
-
-// MARK: - FeedSearchBar
-
-struct FeedSearchBar: View {
-  @State var colorModel: FeedSearchBarColorModel = .init()
-  @State var isNeedBackButton = false
-  @FocusState private var isFocused: Bool
-  @Binding var searchText: String
-  @Binding var isSearching: Bool
-  var cancelTapAction: DismissAction
-
-  var body: some View {
-    ZStack {
-      HStack(spacing: 0) {
-        TextField("", text: $searchText, prompt: Text("Search").foregroundColor(Color.Disable_Placeholder_Dark))
-          .focused($isFocused)
-          .padding(.horizontal, 34)
-          .frame(height: UIScreen.getHeight(28))
-          .foregroundStyle(colorModel.textColor)
-          .fontSystem(fontDesignSystem: .body1)
-          .background(colorModel.bgColor)
-          .cornerRadius(10)
-          .onTapGesture {
-            isFocused = true
-            withAnimation {
-              isSearching = true
-            }
-          }
-          .onSubmit {
-            withAnimation {
-              isSearching = false
-              isFocused = false
-            }
-          }
-          .overlay(alignment: .leading) {
-            Image(systemName: "magnifyingglass")
-              .foregroundStyle(colorModel.placeholderColor)
-              .font(.system(size: 16))
-              .hLeading()
-              .padding(.leading, 8)
-          }
-          .overlay(alignment: .trailing) {
-            if isSearching, !searchText.isEmpty {
-              Button(action: {
-                searchText = ""
-              }) {
-                Image(systemName: "xmark.circle.fill")
-                  .foregroundColor(colorModel.placeholderColor)
-              }
-              .hTrailing()
-              .padding(.trailing, 8)
-            }
-          }
-        if !isNeedBackButton {
-          Text(CommonWords().cancel)
-            .foregroundStyle(colorModel.cancelButtonColor)
-            .fontSystem(fontDesignSystem: .body1)
-            .padding(.leading, 16)
-            .contentShape(Rectangle())
-            .onTapGesture {
-              UIApplication.shared.endEditing()
-              searchText = ""
-              withAnimation {
-                isSearching = false
-              }
-              if isNeedBackButton {
-                return
-              }
-              cancelTapAction()
-            }
-        }
-      }
-    }
-    .onDisappear {
-      UIApplication.shared.endEditing()
-    }
-  }
-}
-
-// MARK: - FeedSearchBarColorModel
-
-class FeedSearchBarColorModel {
-  var bgColor: Color = .init(hex: 0x767680, opacity: 0.24)
-  var placeholderColor: Color = .Disable_Placeholder_Dark
-  var cancelButtonColor: Color = .info
-  var textColor: Color = .LabelColor_Primary
-}
