+++ Whistle/Utilities/Extensions/Extension+Int.swift
@@ -14,4 +14,17 @@ extension Int {
 
     return "\(m_string):\(s_string)"
   }
+
+  var roundedWithAbbreviations: String {
+    let number = Double(self)
+    let thousand = number / 1000
+    let million = number / 1000000
+    if million >= 1.0 {
+      return "\(round(million*10)/10)M"
+    } else if thousand >= 1.0 {
+      return "\(round(thousand*10)/10)K"
+    } else {
+      return "\(self)"
+    }
+  }
 }
